[DEV-154] rename: todo-modal 디렉토리 components로 이동
